refactor: refactoring and renaming files
