+++ one_name_shortener.py
